requirements.txt
Django==2.2.10
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
